requirements.txt
absl-py==2.4.0
altair==5.5.0
annotated-types==0.7.0
appdirs==1.4.4
asgiref==3.11.1
astunparse==1.6.3
attrs==25.4.0
autograd==1.8.0
autoray==0.8.2
bcrypt==5.0.0
bleach==6.2.0
blinker==1.9.0
cachetools==6.2.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
diastatic-malt==2.15.2
dill==0.4.1
Django==5.2.11
filelock==3.20.0
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
flatbuffers==25.12.19
fonttools==4.62.1
fsspec==2025.9.0
gast==0.4.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.48.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==3.2.4
grpcio==1.74.0
gunicorn==25.1.0
h5py==3.15.1
huggingface-hub==0.35.3
ibm-cloud-sdk-core==3.24.4
ibm-platform-services==0.74.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kaggle==1.7.4.5
keras==2.13.1
kiwisolver==1.5.0
libclang==18.1.1
lxml==6.0.2
Markdown==3.10.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
mysql-connector-python==9.5.0
namex==0.1.0
narwhals==2.12.0
networkx==3.5
nltk==3.9.2
numpy<2
oauthlib==3.3.1
opencv-python==4.13.0.92
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
pandas==2.3.3
pdfminer.six==20251107
pdfplumber==0.11.8
pennylane==0.44.1
PennyLane-qiskit==0.44.1
pennylane_lightning==0.44.0
pillow==12.0.0
protobuf==3.20.3
psutil==7.2.2
pyarrow==21.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
PyJWT==2.12.1
PyMuPDF==1.26.6
pyparsing==3.3.2
PyPDF2==3.0.1
pypdfium2==5.0.0
pyspnego==0.12.1
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.2.1
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.3
qiskit==2.3.0
qiskit-aer==0.17.2
qiskit-ibm-runtime==0.45.1
referencing==0.37.0
regex==2025.9.18
requests==2.32.5
requests-oauthlib==2.0.0
requests_ntlm==1.3.0
rich==14.3.2
rpds-py==0.29.0
rsa==4.9.1
rustworkx==0.17.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy-openblas32==0.3.31.22.1
seaborn==0.13.2
semantic-version==2.10.0
sentence-transformers==5.1.1
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.43
sqlparse==0.5.5
sspilib==0.5.0
stevedore==5.7.0
streamlit==1.51.0
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.3.0
text-unidecode==1.3
threadpoolctl==3.6.0
tokenizers==0.22.1
toml==0.10.2
tomlkit==0.14.0
torch==2.10.0
torchvision==0.25.0
tornado==6.5.2
tqdm==4.67.1
transformers==4.57.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
webencodings==0.5.1
Werkzeug==3.1.3
whitenoise==6.12.0
wrapt==2.1.1
django-cors-headers==4.7.0
